remove dummydropdown component
